move create_file to utils
